Split cond, feature and expression to expression package
