+++ querio/ml/__init__.py
@@ -1,2 +1,1 @@
 from .model import Model
-from .feature import Feature
